Create Ddr::Datastreams::MetadataDatastream superclass.
Move selected methods from DescriptiveMetadataDatastream to MetadataDatastream.
Refactor #term_names class methods.
Closes #141
