+++ lib/ddr/datastreams.rb
@@ -23,6 +23,7 @@ module Ddr
 
     CHECKSUM_TYPES = [ CHECKSUM_TYPE_MD5, CHECKSUM_TYPE_SHA1, CHECKSUM_TYPE_SHA256, CHECKSUM_TYPE_SHA384, CHECKSUM_TYPE_SHA512 ]
 
+    autoload :MetadataDatastream
     autoload :AdminMetadataDatastream
     autoload :DescriptiveMetadataDatastream
     #autoload :PreservationMetadataDatastream
